+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-08-10
+### Added
+- Added Arm64EC, LoongArch32/64, NVPTX/64, S390X, and SPARC/64 architectures
+- Added `Feature::can_detect_at_runtime` to report runtime detection availability
+- Added architecture name and family queries
+
+### Changed
+- Split MIPS, PowerPC, RISC-V, WebAssembly, and x86 into 32- and 64-bit variants
+- Made `Architecture` non-exhaustive
+- Renamed `CURRENT_TARGET` to `BUILD_TARGET`
+- Generated feature implications from rustc's target-feature graph
+
+### Fixed
+- Improved `suggested_simd_width`
+
 ## [0.1.6] - 2024-03-15
 ### Changed
 - Updated feature list
@@ -35,7 +50,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/calebzulawski/target-features/compare/0.1.6...HEAD
+[Unreleased]: https://github.com/calebzulawski/target-features/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/calebzulawski/target-features/compare/0.1.6...0.2.0
 [0.1.6]: https://github.com/calebzulawski/target-features/compare/0.1.5...0.1.6
 [0.1.5]: https://github.com/calebzulawski/target-features/compare/0.1.4...0.1.5
 [0.1.4]: https://github.com/calebzulawski/target-features/compare/0.1.3...0.1.4